+++ test/generators/migration_generator_test.rb
@@ -53,4 +53,18 @@ class MigrationGeneratorTest < Rails::Generators::TestCase
       assert_migration "db/migrate/create_rodauth_email_auth.rb", /default: OpenStruct\.new\(quoted_id: "CURRENT_TIMESTAMP"\)/
     end
   end
+
+  test "no features" do
+    output = run_generator %w[]
+
+    assert_equal "No features specified!\n", output
+    assert_no_file "db/migrate/create_rodauth_.rb"
+  end
+
+  test "invalid features" do
+    output = run_generator %w[sms_codes active_session totp]
+
+    assert_equal "No available migration for feature(s): active_session, totp\n", output
+    assert_no_file "db/migrate/create_rodauth_otp.rb"
+  end
 end